refactor to fix build
