+++ kin-sdk-core/src/androidTest/java/kin/sdk/core/util/KinConverterTest.java
@@ -1,4 +1,4 @@
-package kin.sdk.core.ethereum;
+package kin.sdk.core.util;
 
 import static org.junit.Assert.assertEquals;
